requirements.txt
fastapi
uvicorn
transformers
torch
pillow
pymilvus
colpali